+++ packages/image_picker/image_picker_platform_interface/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 2.2.1
+
+* Update cache system to save list of images and retrieve them when it is needed. On Android, 
+`List<PickedFile> files` property added to `LostData` and `List<XFile> files` property added to 
+`LostDataResponse` in the case multiple picked images were recovered.
+* Refactor `retrieveLostData` and `getLostData` methods for new cache system.
+
 ## 2.2.0
 
 * Added new methods that return `XFile` (from `package:cross_file`)